IBP-5093: Move list creation classes to shared/list-creation
