+++ src/main/jhipster/src/main/webapp/app/shared/list-creation/model/germplasm-list.ts
@@ -1,8 +1,7 @@
-import { SearchComposite } from './search-composite';
-import { GermplasmSearchRequest } from '../../entities/germplasm/germplasm-search-request.model';
-import { ListEntry, ListModel } from '../list-builder/model/list.model';
+import { SearchComposite } from '../../model/search-composite';
+import { GermplasmSearchRequest } from '../../../entities/germplasm/germplasm-search-request.model';
+import { ListEntry, ListModel } from '../../list-builder/model/list.model';
 
-// TODO move to shared/list-creation/model
 export class GermplasmList extends ListModel {
     constructor(
         public searchComposite?: SearchComposite<GermplasmSearchRequest, number>
